Separate Blog component from page route component
